fix(cli): Don't load HathorSettings before entering command main method
